Refactoring: Extracted command-line driven performance test to separate class
